+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+
+## 31.11.0 - 2018-08-01
 ### Added
 - [Feature] Added DiscloseTable component
 - [Patch] Added width prop for activator of SelectDropdown